Update subdomains.py
This function contains a dictionary with information regarding various subdomains. Its main purpose is to store latitude and longitude information for the subdomain. It is most useful when WRF plotting scripts calls the function get_domain('name_of_subdomain') and extracts the latitude and longitude bounds.

See example at bottom